requirements.txt
requests
numpy
nose
pandas
ipython
ipdb
pyyaml
pytest-sugar